backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic[email]==2.8.2
openpyxl==3.1.5
google-generativeai==0.8.3
python-dotenv==1.0.1
httpx==0.27.2